backend/requirements.txt
# Strands Agents + Bedrock
strands-agents
strands-agents-tools
boto3>=1.35.0

# Web server
fastapi
uvicorn[standard]
python-multipart

# Weather API
httpx

# CORS
python-dotenv
